음성인식 TEXT Callback Return !
-  AR, 신규 데이터 생성 요청 메소드 추가 (PalaceCore :: requestNewAugmentedItem)

+++ mobile/src/main/java/kr/poturns/virtualpalace/controller/data/ResourceItem.java
@@ -4,7 +4,7 @@ package kr.poturns.virtualpalace.controller.data;
  *
  */
 public class ResourceItem {
-    public int _id;
+    public long _id;
 
     public String title;
